requirements.txt
Flask==1.1.1
line-bot-sdk==1.19.0
gunicorn==20.0.4